LayoutTests/fast/frames/frame-element-name-expected.txt



--------
Frame: 'left'
--------
PASS escape(window.frameElement.name) is "left"
PASS escape(window.name) is "left"


--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
PASS escape(window.frameElement.name) is ""
PASS escape(window.name) is ""
